Refactor camera-related classes into a dedicated 'camera' package
