+++ peekaboo-ui/src/androidMain/kotlin/com/preat/peekaboo/ui/camera/PeekabooCamera.android.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.preat.peekaboo.ui
+package com.preat.peekaboo.ui.camera
 
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
