+++ peekaboo-ui/src/commonMain/kotlin/com/preat/peekaboo/ui/camera/PeekabooCamera.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.preat.peekaboo.ui
+package com.preat.peekaboo.ui.camera
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
