+++ peekaboo-ui/src/iosMain/kotlin/com/preat/peekaboo/ui/camera/PeekabooCamera.ios.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.preat.peekaboo.ui
+package com.preat.peekaboo.ui.camera
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
